Revert "Update"
This reverts commit 018fee455e8e15c6d02655012257c5eed98f47ce.

+++ include/code/main.h
@@ -12,7 +12,7 @@ extern PreNmiBuff* gAppNmiBufferPtr;
 extern Scheduler gScheduler;
 extern PadMgr gPadMgr;
 extern IrqMgr gIrqMgr;
-extern uintptr_t gSegments[NUM_SEGMENTS];
+extern u32 gSegments[NUM_SEGMENTS];
 extern OSThread sGraphThread;
 STACK(sGraphStack, 0x1800);
 STACK(sSchedStack, 0x600);
